Add a standalone recipe for Qt 4.8 Lighthouse
For now it uses a local Qt 4.8 source checkout. Patches still pending inclusion.
Note that this recipe is NOT based on the previous versions which use a separate set of recipes
for the native tools (qmake, moc, etc.).
Instead, we use a simple qmake builder class which builds everything in one go and which is much simpler to maintain.