Remove explicit OpenGL include/lib dirs from the makespec
The includes are in /usr/include and the libs in /usr/lib anyway, so this isn't needed.
And this actually causes OpenGL support to be badly detected (only ES1 support), as
the QT_SYSROOT variable isn't set at that point, causing the test to search in the
system /usr/include instead of our staging dir.